Got the mesh rendering correctly
Fixed a bug where the indices of the mesh were in the wrong order so its winding order normal was pointing downwards.

Also fixed a bug where the end X triangles index was getting wrapped back to the start of the mesh X position, causing it to look like there were actually two meshes because it stretched all the way back to the start from one end.