Implemented an object that renders using a geometry shader
The geometry object renders a 4 point star from one vertex.

Renamed all the shaders so they are automatically organised in windows.
(this caused me to get stuck on an error for like 2 hours where it couldn't find a uniform in the geometry fragment shader, this was because there was an extra space at the end of the filename :)))))))))))))))))))))))))))))))))))