Added a third person camera
The third person camera gets attached to the m_cube in the game manager.

Added input to move m_cube around using WASD.

Fixed a crash that would happen if m_cube went out of bounds of the terrain.

Added a toggle to switch between free camera mode, and the third person mode.